+++ Back-End/src/main/java/br/com/wsb/DonJose/model/Categoria.java
@@ -19,18 +19,18 @@ public class Categoria implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id 
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	
 	private String nome;
 	 
 	@ManyToMany(mappedBy="categorias")
 	private List<Produto> produtos = new ArrayList<>();
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -56,7 +56,7 @@ public class Categoria implements Serializable {
 		super();
 	}
 
-	public Categoria(long id,String nome) {
+	public Categoria(Integer id,String nome) {
 		super();
 		this.id = id;
 		this.nome = nome;
