+++ Back-End/src/main/java/br/com/wsb/DonJose/service/CategoriaService.java
@@ -22,7 +22,7 @@ public class CategoriaService {
 	@Autowired
 	private CategoriaRepository Catrepository;
 
-	public Categoria find(long id) {
+	public Categoria find(Integer id) {
 		Optional<Categoria> obj = Catrepository.findById(id);
 		return obj.orElseThrow(() -> new ObjectNotFoundException(
 				"Objeto não encontrado! Id: " + id + ", Tipo: " + Categoria.class.getName()));
@@ -39,7 +39,7 @@ public class CategoriaService {
 		return Catrepository.save(newObj);
 	}
 
-	public void delete(long id) {
+	public void delete(Integer id) {
 		find(id);
 		try {
 			Catrepository.deleteById(id);
